Refactor code of plugin sharing
