+++ packages/gitbook-plugin-sharing/package.json
