+++ packages/gitbook-plugin-sharing/src/shapes/site.js
@@ -1,0 +1,13 @@
+const {
+    string,
+    func,
+    shape
+} = require('gitbook-core').React.PropTypes;
+
+const siteShape = shape({
+    label: string.isRequired,
+    icon: string.isRequired,
+    onShare: func.isRequired
+});
+
+module.exports = siteShape;
