+++ packages/gitbook/package.json
